Feature/test product

+++ src/main/java/mori/task/interview/tariffComparison/model/PackagedTarrif.java
@@ -4,15 +4,15 @@ import java.math.BigDecimal;
 
 public class PackagedTarrif extends Product {
 
-    public PackagedTarrif(int consumption){
-        super("Packaged Tariff",consumption);
+    public PackagedTarrif(double consumption) {
+        super("Packaged Tariff", consumption);
     }
 
     @Override
-    public BigDecimal calculationModel(int consumption) {
-        if(consumption <= 4000 ){
-            return BigDecimal.valueOf(800);
-        }else{
+    public BigDecimal calculationModel(double consumption) {
+        if (consumption <= 4000) {
+            return BigDecimal.valueOf(800.0);
+        } else {
             double additionalConsumptionCosts = consumption - 4000;
             return BigDecimal.valueOf(800 + additionalConsumptionCosts * 0.30);
         }
